mainframe: fix comparison between signed and unsigned integer

+++ src/mainframe.cpp
@@ -1558,7 +1558,7 @@ void CMainFrame::CreateFavoritesMenu()
 	{
 		CSite& site = *it;
 		GtkWidget* fav_item;
-		const int fav_string_len = 20;
+		const size_t fav_string_len = 20;
 		fav_item = gtk_image_menu_item_new_with_label(
 			site.m_Name.length() <= fav_string_len ?
 			site.m_Name.c_str() :
